+++ examples/reflex/rxconfig.py
@@ -1,4 +1,5 @@
 import reflex as rx
+
 import reflex_xy
 
 config = rx.Config(